injected NotificationManagerCompat
