--- core/notification/src/main/java/edu/stanford/spezi/core/notification/notifier/NoOpNotifier.kt
@@ -1,10 +1,0 @@
-package edu.stanford.spezi.core.notification.notifier
-
-import javax.inject.Inject
-
-/**
- * A [NotificationNotifier] that does nothing. Useful for testing.
- */
-internal class NoOpNotifier @Inject constructor() : NotificationNotifier {
-    override fun sendNotification(firebaseMessage: FirebaseMessage) = Unit
-}
